Change redis key to use SeriesInstanceUID only
